requirements.txt
asgiref==3.4.1

Django==3.2.25
djangorestframework==3.15.1
importlib-resources==5.4.0
pytz==2025.2
sqlparse==0.4.4
typing-extensions==4.1.1
zipp==3.6.0
